+++ sdl3-image-sys/ChangeLog.md
@@ -1,6 +1,7 @@
 # Changes
 
     - Enable `use-pkg-config` and `use-vcpkg` features by default
+    - Fix Android builds being built for the wrong arch (Thanks to Archee)
 
 - 0.6.3
     - Update SDL_image to 3.4.4