+++ sdl3-sys/ChangeLog.md
@@ -4,6 +4,7 @@
     - Update SDL to 3.4.10
     - Enable `use-pkg-config` and `use-vcpkg` features by default
     - Replace use of deprecated `cdecl` and `stdcall` ABIs with equivalent non-deprecated ones
+    - Fix Android builds being built for the wrong arch (Thanks to Archee)
 
 - 0.6.5:
     - Update SDL to 3.4.8 (hotfix)